requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Machine Learning and Image Processing
torch==2.2.0
torchvision==0.17.0
pillow==10.1.0
numpy==1.26.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development and utilities
python-dotenv==1.0.0
requests==2.31.0

# Optional: For production deployment
gunicorn==21.2.0 